requirements.txt
fastapi
uvicorn
requests
SQLAlchemy
python-dotenv
# databases[postgresql]
# alembic
# pydantic[email]
# passlib[bcrypt]
# fastapi-mail
# python-jose[cryptography]
# python-multipart
PyMySQL
# pip install -r requirements.txt
# py -m venv env
# source ./Scripts/activate